==> twitter_pull_and_processing/1_lets_get_those_tweets_copy.ipynb <==
import tweepy, sys, os, logging, json

# Twitter API using Tweepy

Set up activity log file:

logging.basicConfig(filename='twitter_error_log.log',filemode='w', level=logging.ERROR)

Load consumer key and secret securely:

consumer_key = ''
consumer_secret = ''

Initialize API

auth = tweepy.AppAuthHandler(consumer_key, consumer_secret)
 
api = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True,
                 parser=tweepy.parsers.JSONParser()
                )
 
if (not api):
    print ("Can't Authenticate")
    sys.exit(-1)

Define twitter searches

# Twitter API search terms
climate_search = ('(global AND warming) OR #climate')
victoria_search = ('victoria secret')
news_search = ('(trump OR clinton) AND (health OR obamacare)')
facebook_search = ('facebook AND (data OR email OR user OR emails OR documents OR confidential)')
family_sep_search = ('family separation')
basic_income_search = ('(universal AND basic AND income) OR (basic AND income)')
democracy_search = ('democracy')

searchQuery = democracy_search

### Running the queries for all the tweets on a certain topic

maxTweets = 1000000 # Some arbitrary large number
tweetsPerQry = 100  # this is the max the API permits

# If results from a specific ID onwards are reqd, set since_id to that ID.
# else default to no lower limit, go as far back as API allows
sinceId = None

# If results only below a specific ID are reqd, set max_id to that ID.
# else default to no upper limit, start from the most recent tweet matching the search query.
max_id = -1

tweetCount = 0
tweetFiles = 0
msg =  "Downloading max {0} tweets".format(maxTweets)
logging.error(msg)
tweets = []
while 4000*tweetFiles + tweetCount < maxTweets:
    try:
        if (max_id <= 0):
            if (not sinceId):
                new_tweets = api.search(q=searchQuery, count=tweetsPerQry)
            else:
                new_tweets = api.search(q=searchQuery, count=tweetsPerQry,
                                        since_id=sinceId)
        else:
            if (not sinceId):
                new_tweets = api.search(q=searchQuery, count=tweetsPerQry,
                                        max_id=str(max_id - 1))
            else:
                new_tweets = api.search(q=searchQuery, count=tweetsPerQry,
                                        max_id=str(max_id - 1),
                                        since_id=sinceId)
                
        new_tweets = new_tweets['statuses']
        if len(new_tweets) == 0:
            msg = 'No tweets found'
            logging.error(msg)
            break
        
        tweets.extend(new_tweets)
        tweetCount += len(new_tweets)
        msg = "Downloaded {0} tweets".format(tweetCount + tweetFiles*4000)
        logging.error(msg)
        max_id = new_tweets[-1]['id']
        
        if tweetCount > 4000:
            with open('tweets'+str(tweetFiles)+'.json', 'w') as outfile:  
                json.dump(tweets, outfile)
            msg = "JSON file saved"
            logging.error(msg)
            tweetFiles += 1
            tweetCount = 0
            del tweets[:]
                
    except tweepy.TweepError as e:
        msg = 'Query failed when max_id equaled {0}: {1}'.format(max_id, e)
        logging.error(msg)

if tweetCount > 0:
    with open('tweets'+str(tweetFiles)+'.json', 'w') as outfile:  
        json.dump(tweets, outfile)
    msg = "JSON file saved"
    logging.error(msg)
    tweetFiles += 1
    tweetCount = 0

# Check the result of one file

with open('tweets0.json') as json_file:  
    test = json.load(json_file)

print(len(test))
test[0]

# Check rate limit usage

import requests
from application_only_auth import Client
client = Client(consumer_key, consumer_secret)

response = client.request(
    "https://api.twitter.com/1.1/application/rate_limit_status.json?resources=help,users,search,statuses")

response
