==> twitter_pull_and_processing/2_reading_JSON_files.ipynb <==
# Processing Twitter data

### A look at a single file

First we load a single file with 4000 tweets into a dataframe to check for nested dictionaries

import pandas as pd
import numpy as np

df = pd.read_json('tweets0.json')

First we define features that will go straight into the dataframe without any unpacking requiered

Sequentially unpack nested lists applying **pd.Series**

df_flat = pd.DataFrame()
# id and date
df_flat['id'] = df['id']
df_flat['date'] = df['created_at']
# user info
user_info = df['user'].apply(pd.Series)
df_flat['user'] = user_info.name
df_flat['user_name'] = user_info.screen_name
df_flat['place'] = user_info.location
# tweet info
df_flat['text'] = df['text']
df_flat['hashtags'] = df['entities'].apply(pd.Series).hashtags
df_flat['mentions'] = df['entities'].apply(pd.Series).user_mentions
df_flat['favorite_count'] = df['favorite_count']
df_flat['retweet_count'] = df['retweet_count']
# retweet user info
retweet_info = df['retweeted_status'].apply(pd.Series)
df_flat['retweet_user'] = retweet_info.user.apply(pd.Series).name
df_flat['retweet_user_name'] = retweet_info.user.apply(pd.Series).screen_name
# quoted user info
quoted_info = df['quoted_status'].apply(pd.Series)
df_flat['quoted_user'] = quoted_info.user.apply(pd.Series).name
df_flat['quoted_user_name'] = quoted_info.user.apply(pd.Series).screen_name
# reply user info
df_flat['reply_user_name'] = df['in_reply_to_screen_name']
df_flat.head()

df_flat.to_csv('flat_tweets.csv')

### Looping through all files

import os, json

json_files = [pos_json for pos_json in os.listdir() if pos_json.endswith('.json')]
print(json_files[:3])

df = pd.DataFrame()

for file in json_files:
    data = pd.read_json(file)
    df = df.append(data)

df.shape

df_flat = pd.DataFrame()
# id and date
df_flat['id'] = df['id']
df_flat['date'] = df['created_at']
# user info
user_info = df['user'].apply(pd.Series)
df_flat['user'] = user_info.name
df_flat['user_name'] = user_info.screen_name
df_flat['place'] = user_info.location
# tweet info
df_flat['text'] = df['text']
df_flat['hashtags'] = df['entities'].apply(pd.Series).hashtags
df_flat['mentions'] = df['entities'].apply(pd.Series).user_mentions
df_flat['favorite_count'] = df['favorite_count']
df_flat['retweet_count'] = df['retweet_count']
df_flat['language'] = df['lang']
# retweet user info
retweet_info = df['retweeted_status'].apply(pd.Series)
df_flat['retweet_user'] = retweet_info.user.apply(pd.Series).name
df_flat['retweet_user_name'] = retweet_info.user.apply(pd.Series).screen_name
# quoted user info
quoted_info = df['quoted_status'].apply(pd.Series)
df_flat['quoted_user'] = quoted_info.user.apply(pd.Series).name
df_flat['quoted_user_name'] = quoted_info.user.apply(pd.Series).screen_name
# reply user info
df_flat['reply_user_name'] = df['in_reply_to_screen_name']
df_flat.head()

df_flat.columns

df_flat.to_csv('all_tweets.csv')

# Exploring the dataset

# unique users
len(df_flat.user.unique())

aggregation = {
    'retweet_count': 'sum',
    'id': 'count'
}

A look at the users with the most retweets

df_flat[df_flat.retweet_user.isna()].groupby(by = 'user_name').agg(aggregation).sort_values('retweet_count',ascending=False).head()

A look at the users that retweeted the most

print(min(df_flat.date))
print(max(df_flat.date))

df_flat[df_flat.retweet_user.notna()].groupby(by = 'user_name').agg({'id':'count'}).sort_values('id',ascending=False).head()
